refactor: fe - 조건문 util로 분리 #61
