+++ frontend/src/utils/isSelectedCursor.ts
@@ -1,0 +1,3 @@
+export function isSelectedCursor(cursor: number, type: number): boolean {
+	return cursor === type;
+}
